remove errors

+++ migrations/000001_create_users.up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
     id          INT AUTO_INCREMENT PRIMARY KEY,
-    telegram_id INT,
+    telegram_id BIGINT,
     created_at  DATETIME DEFAULT CURRENT_TIMESTAMP
 );
